moving ofac to indexer folder and minor changes there
