+++ src/indexer/index.ts
@@ -8,14 +8,11 @@ import TransferRepository from "./repository/transfer"
 import ExecutionRepository from "./repository/execution"
 import FeeRepository from "./repository/fee"
 import ResourceRepository from "./repository/resource"
-import { OfacComplianceService } from "../services/ofac.service"
+import { OfacComplianceService } from "./services/evmIndexer/ofac"
 
 async function main(): Promise<void> {
   const sharedConfig = await getSharedConfig(process.env.SHARED_CONFIG_URL!)
-  const ofacComplianceService = new OfacComplianceService(
-    process.env.CHAIN_ANALYSIS_URL!,
-    process.env.CHAIN_ANALYSIS_API_KEY!
-  )
+  const ofacComplianceService = new OfacComplianceService(process.env.CHAIN_ANALYSIS_URL!, process.env.CHAIN_ANALYSIS_API_KEY!)
 
   const domainRepository = new DomainRepository()
   const depositRepository = new DepositRepository()
@@ -48,7 +45,16 @@ async function main(): Promise<void> {
       }
     } else if (domain.type == DomainTypes.EVM) {
       try {
-        const evmIndexer = new EvmIndexer(domain, rpcURL, domainRepository, depositRepository, transferRepository, executionRepository, feeRepository, ofacComplianceService)
+        const evmIndexer = new EvmIndexer(
+          domain,
+          rpcURL,
+          domainRepository,
+          depositRepository,
+          transferRepository,
+          executionRepository,
+          feeRepository,
+          ofacComplianceService,
+        )
         await evmIndexer.listenToEvents()
       } catch (err) {
         logger.error(`error on domain: ${domain.id}... skipping`)
