+++ src/indexer/services/evmIndexer/evmIndexer.ts
@@ -11,7 +11,7 @@ import { logger } from "../../../utils/logger"
 import { getLogs } from "./evmfilter"
 import { DecodedLogs } from "./evmTypes"
 import { decodeLogs } from "./evmEventParser"
-import { OfacComplianceService } from "../../../services/ofac.service"
+import { OfacComplianceService } from "./ofac"
 
 export class EvmIndexer {
   private provider: ethers.JsonRpcProvider
@@ -34,7 +34,7 @@ export class EvmIndexer {
     transferRepository: TransferRepository,
     executionRepository: ExecutionRepository,
     feeRepository: FeeRepository,
-    ofacComplianceService: OfacComplianceService
+    ofacComplianceService: OfacComplianceService,
   ) {
     this.provider = new ethers.JsonRpcProvider(rpcURL)
     this.domainRepository = domainRepository
@@ -126,7 +126,9 @@ export class EvmIndexer {
       const transferMap = new Map<string, string>()
 
       await Promise.all(
-        decodedLogs.deposit.map(async decodedLog => saveDepositLogs(decodedLog, this.transferRepository, this.depositRepository, transferMap, this.ofacComplianceService)),
+        decodedLogs.deposit.map(async decodedLog =>
+          saveDepositLogs(decodedLog, this.transferRepository, this.depositRepository, transferMap, this.ofacComplianceService),
+        ),
       )
 
       await Promise.all(decodedLogs.feeCollected.map(async fee => saveFeeLogs(fee, transferMap, this.feeRepository)))
